long(0) -> 0
There is no need to cast 0 to 'long' even if we want to compare it to long
numbers in modern pythons.

+++ offlineimap/folder/Maildir.py
@@ -40,7 +40,7 @@ re_uidmatch = re.compile(',U=(\d+)')
 re_timestampmatch = re.compile('(\d+)');
 
 timeseq = 0
-lasttime = long(0)
+lasttime = 0
 timelock = Lock()
 
 def gettimeseq():
